+++ packages/react-utils/src/index.ts
@@ -6,3 +6,4 @@ export * from "./use-presence";
 export type * from "./types";
 
 export * from "./is";
+export * from "./use-previous";