+++ packages/react/src/scrollable/scrollable.css.ts
@@ -1,0 +1,12 @@
+import { style } from "@vanilla-extract/css";
+
+const wrapper = style({
+});
+const content = style({
+
+});
+
+export const styles = {
+  wrapper,
+  content,
+} as const;